requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- DNABERT2 --------- #
transformers==4.34.0
datasets==2.16.1  # For handling dataset pipelines
scikit-learn>=1.2.2  # For evaluation metrics

# --------- others --------- #
rootutils
pre-commit
rich
pytest
